Now give

+++ frontAngTest/src/app/body/body.component.ts
@@ -49,22 +49,23 @@ export class BodyComponent implements OnInit {
   }
 
   deleteDevice(device){
-    alert(device.name + " has been remove");
-    const jwt = window.localStorage.getItem('jwt')
-    const headers = new HttpHeaders({
-      'Content-Type': 'application/json',
-      'Authorization': jwt
-    })
-    let newStatus=null
-    const host = (window.location.href.split("/")[2]).split(":")[0]
-    let url = null
-    url = "http://" + host + ":8000/device/" + device.name
-    this.http.delete(url, { headers: headers }).subscribe( data =>
-    {
-      if(data!=null){
-        this.getDevicesList()
-      }
-    })
+    if (confirm("Delete " + device + "?")) {
+      const jwt = window.localStorage.getItem('jwt')
+      const headers = new HttpHeaders({
+        'Content-Type': 'application/json',
+        'Authorization': jwt
+      })
+      let newStatus=null
+      const host = (window.location.href.split("/")[2]).split(":")[0]
+      let url = null
+      url = "http://" + host + ":8000/device/" + device.name
+      this.http.delete(url, { headers: headers }).subscribe( data =>
+      {
+        if(data!=null){
+          this.getDevicesList()
+        }
+      })
+    }
   }
 
   getDifference(startTime, finishTime){
